Moved toMsg to other utils
