+++ src/client/logic/api.ts
@@ -1,7 +1,7 @@
 import * as _ from 'lodash';
 import { vkConnect } from '../external';
 import { ILaunchParams, Methods, QueryParams, RequestParams, ResponseType, IResponse, IAdminGroupConfig, RequestType, GetRequestType, GetRequestRoute, IGroupRequirement, TRepostInfo, IUserStatus, IPromocode } from '../../common/types';
-import { toMsg } from '../../common/errors';
+import { toMsg } from '../../common/utils';
 import { vkAuthHeaderName, vkApiAuthHeaderName } from '../../common/security';
 import { isOk } from '../../common/utils';
