+++ src/server/router-ex.ts
@@ -2,7 +2,7 @@ import * as express from 'express';
 import * as Joi from '@hapi/joi';
 
 import { RequestType, Methods, GetRequestType, GetRequestRoute, IResponse, ResponseType, QueryParams, RequestParams, GetQueryParamsSchema } from "../common/types";
-import { toMsg } from '../common/errors';
+import { toMsg } from '../common/utils';
 
 type RouterMethodTypes = "get" | "put";
 interface IRouterMapper {
