+++ src/server/security.ts
@@ -4,10 +4,11 @@ import * as _ from 'lodash';
 import * as Joi  from '@hapi/joi';
 
 import { Request, NextFunction, Response, RequestHandler} from 'express';
-import { toMsg, sendError } from '../common/errors';
+import { toMsg } from '../common/utils';
 import { vkAuthHeaderName, vkApiAuthHeaderName } from '../common/security';
 import { IVkParams, VkViewerRole } from '../common/types';
 import { isOk } from '../common/utils';
+import { sendError } from '../common/errors';
 
 
 declare global {
